Split test classes into separate files
